requirements.txt
tensorflow>=2.4.0
tensorflow-addons>=0.13.0
numpy>=1.19.5
pandas>=1.2.0
matplotlib>=3.3.4
scikit-image>=0.18.1
scipy>=1.6.0
patchify>=0.2.3
tqdm>=4.59.0
pillow>=8.1.0
efficientnet>=1.1.1
tensorboard>=2.4.0
h5py>=3.1.0 